Added class 25/09

+++ Banco/src/banco/CDB.java
@@ -5,6 +5,8 @@
  */
 package banco;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author eduardo.schneider
@@ -12,9 +14,9 @@ package banco;
 public class CDB {
     private int id;
     private String nome;
-    private double saldo;
+    private BigDecimal saldo;
 
-    public CDB(int id, String nome, double saldo) {
+    public CDB(int id, String nome, BigDecimal saldo) {
         this.id = id;
         this.nome = nome;
         this.saldo = saldo;
@@ -36,11 +38,11 @@ public class CDB {
         this.nome = nome;
     }
 
-    public double getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(double saldo) {
+    public void setSaldo(BigDecimal saldo) {
         this.saldo = saldo;
     }
 }
